requirements.txt
strands-agents
PyYAML
numpy
openai
anthropic
python-dotenv
pylatexenc
sympy